Parcel things to separate files
We still pull them in from swt/full.rb, but this allows for clients to
be pickier about exactly which parts they import if they'd like. They
can grab swt/minimal, and then whatever individual parts they want.
