+++ lib/swt/swt_bot_extensions.rb
@@ -1,4 +1,3 @@
-
 module SwtBotExtensions
   def c_tab_folder
     c_tab_item.widget.parent
